YaraAutoTestWeb/YARA_TEST/requirements.txt
alembic==1.1.0
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
beautifulsoup4==4.8.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
configparser==3.7.4
constantly==15.1.0
contextvars==2.4
cryptography==2.7
cssselect==1.0.3
devconfig==0.4.6
et-xmlfile==1.0.1
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
future==0.17.1
gevent==1.4.0
greenlet==0.4.15
hyperlink==19.0.0
idna==2.8
immutables==0.9
incremental==17.5.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
lxml==4.4.0
Mako==1.1.0
MarkupSafe==1.1.1
Naked==0.1.31
openpyxl==2.6.2
openssl-python==0.1.1
parsel==1.5.1
pefile==2019.4.18
pyasn1==0.4.6
pyasn1-modules==0.2.5
pycparser==2.19
pycrypto==2.6.1
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
python-dateutil==2.8.0
python-editor==1.0.4
python-json-logger==0.1.11
PyYAML==5.1.1
queuelib==1.5.0
requests==2.22.0
Scrapy==1.7.2
service-identity==18.1.0
shellescape==3.4.1
six==1.12.0
soupsieve==1.9.2
SQLAlchemy==1.3.8
tornado==6.0.3
Twisted==19.2.1
urllib3==1.25.3
w3lib==1.20.0
Werkzeug==0.15.5
WTForms==2.2.1
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
yara-python==3.8.1
yaratool==0.0.6
zope.interface==4.6.0
